feat(cwm): add config context
